semana05/dia5/requirements.txt
django==3.2
djangorestframework
django-cors-headers
cloudinary
djangorestframework-simplejwt
